+++ src/components/main/DateFormat/index.ts
@@ -1,0 +1,9 @@
+export const formatDate = (date: Date | null): string | null => {
+  if (date) {
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, '0');
+    const day = String(date.getDate()).padStart(2, '0');
+    return `${year}-${month}-${day}`;
+  }
+  return null;
+};
